Avoid local import of the threading module
In the past the threading module was not available on all platforms for
cpython, but this is no-longer the case. (I'm not sure if this is the
reason why threading was imported locally).

Since we don't support very ancient cpython versions and we didn't
import 'threading' locally, due to cyclic import issues, we should
import threading globally.